Go button for view by muscle groups
Force a search for muscle groups. Can be used after searching via name,
which populates the exercise list with the matched exercise search
query. Since the selection can still be on "all groups" and the exercise
list still contain the prevoius search results, this button can be used
to force a search.